hostpid: retrieve the host PID using a bpf program
This patch introduces a bpf program for determining the host PID of the
running process. This will be used in subsequent patches to check
whether the proc file system is actuall the host /proc.

The program runs only once and independently from other sensors. It uses
bogus arguments in the lseek system call to identify the running
process.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>